Share one firing animation hook between the transition nodes, and fold the zoom gate into the visibility check
