+++ .changeset/canvas-firing-animation-budget.md
@@ -2,4 +2,4 @@
 "@hashintel/petrinaut": patch
 ---
 
-A firing only animates where it can be seen: nodes and arcs off the side of the canvas, or drawn at a zoom small enough that a node is a few pixels across, skip the flash. On a thousand-node net a scrub goes from 13 to 44 frames per second and playback from 16 to 39.
+Firing animations run only on the transitions and arcs in view, and not at all when the canvas is zoomed out far enough that a node is a few pixels across.
